Keyframing and Path Display
Using the priority queue (openList), the f values were sorted (sum of the heuristic predictions and distance from start).
The main loop selects the node with the lowest f value from openList. If selected node was the end node, recontructPath was called. Otherwise, add the selected node to closeList (or list of nodes that are invalid).
The recontructPath method highlights the final path from the end node to the start node.