Keyframes & Path highlights
Added a Timeline with a KeyFrame to animate the A* algorithm. The KeyFrame updates the path every 0.1 seconds, making the algorithm run one step at a time.
The Plathform.runLater ensures that the UI is constantly updated to display the path on the grid.
Added the 'highlightPath' method to highlight the path from the starting node to the current node.